fix : 패키지 경로 변경 (example -> destiny)
