+++ review-service/src/main/java/com/destiny/reviewservice/ReviewServiceApplication.java
@@ -1,0 +1,13 @@
+package com.destiny.reviewservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReviewServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ReviewServiceApplication.class, args);
+    }
+
+}
