+++ review-service/src/main/java/com/destiny/reviewservice/infrastructure/client/config/FeignClientConfig.java
@@ -1,0 +1,10 @@
+package com.destiny.reviewservice.infrastructure.client.config;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients(basePackages = "com.destiny.reviewservice")
+public class FeignClientConfig {
+
+}
