+++ review-service/src/main/java/com/destiny/reviewservice/infrastructure/jpa/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.destiny.reviewservice.infrastructure.jpa.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+}
