+++ review-service/src/main/java/com/destiny/reviewservice/infrastructure/security/auth/CustomAccessDeniedHandler.java
@@ -1,0 +1,31 @@
+package com.destiny.reviewservice.infrastructure.security.auth;
+
+import com.destiny.global.code.CommonErrorCode;
+import com.destiny.global.response.ApiResponse;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response,
+        AccessDeniedException accessDeniedException) throws IOException, ServletException {
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+        response.getWriter().print(objectMapper.writeValueAsString(
+            ApiResponse.error(CommonErrorCode.ACCESS_DENIED)
+        ));
+    }
+}
