+++ review-service/src/main/java/com/destiny/reviewservice/infrastructure/security/auth/CustomUserDetailsService.java
@@ -1,0 +1,19 @@
+package com.destiny.reviewservice.infrastructure.security.auth;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CustomUserDetailsService implements UserDetailsService {
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        throw new UsernameNotFoundException("현재 사용하지 않는 인증 방식입니다.");
+    }
+}
