+++ review-service/src/main/java/com/destiny/reviewservice/presentation/dto/request/ReviewUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.destiny.reviewservice.presentation.dto.request;
+
+public record ReviewUpdateRequest(
+    int rating,
+    String content,
+    String imageUrls
+) {
+
+}
