--- review-service/src/main/java/com/example/reviewservice/ReviewServiceApplication.java
@@ -1,13 +1,0 @@
-package com.example.reviewservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReviewServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ReviewServiceApplication.class, args);
-    }
-
-}
