--- review-service/src/main/java/com/example/reviewservice/infrastructure/jpa/config/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.example.reviewservice.infrastructure.jpa.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-}
