--- review-service/src/main/java/com/example/reviewservice/infrastructure/jpa/config/SecurityAuditorAware.java
@@ -1,29 +1,0 @@
-package com.example.reviewservice.infrastructure.jpa.config;
-
-import com.example.reviewservice.infrastructure.security.auth.CustomUserDetails;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class SecurityAuditorAware implements AuditorAware<UUID> {
-
-    @Override
-    public Optional<UUID> getCurrentAuditor() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            return Optional.empty();
-        }
-
-        Object principal = authentication.getPrincipal();
-        if (principal instanceof CustomUserDetails userDetails) {
-            return Optional.of(userDetails.getUserId());
-        }
-
-        return Optional.empty();
-    }
-}
