--- review-service/src/main/java/com/example/reviewservice/infrastructure/security/auth/CustomAuthenticationEntryPoint.java
@@ -1,31 +1,0 @@
-package com.example.reviewservice.infrastructure.security.auth;
-
-import com.destiny.global.code.CommonErrorCode;
-import com.destiny.global.response.ApiResponse;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-        AuthenticationException authException) throws IOException, ServletException {
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        response.setContentType("application/json");
-        response.setCharacterEncoding("UTF-8");
-        response.getWriter().print(objectMapper.writeValueAsString(
-            ApiResponse.error(CommonErrorCode.UNAUTHORIZED)
-        ));
-    }
-}
