--- review-service/src/main/java/com/example/reviewservice/presentation/dto/request/ReviewUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.example.reviewservice.presentation.dto.request;
-
-public record ReviewUpdateRequest(
-    int rating,
-    String content,
-    String imageUrls
-) {
-
-}
