+++ review-service/src/test/java/com/destiny/reviewservice/ReviewServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.destiny.reviewservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReviewServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
