--- review-service/src/test/java/com/example/reviewservice/ReviewServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.reviewservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReviewServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
